Added Mathematica language
As Wolfram Mathematica includes thousands of built-in functions that
should be colored, this language component is approximately as big as
the rest of Prism. For this reason, I will not send a pull request
unless someone come up with an alternative to include this component
without forcing every user of Prism to download it.